[SQL] Use a single WINDOW operator for temporal filters instead of two when possible
