+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/visitors/inner/Simplify.java
@@ -691,6 +691,13 @@ public class Simplify extends ExpressionTranslator {
                             result = right;
                         }
                     }
+                } else if (left instanceof DBSPUnaryExpression unaryLeft &&
+                        right instanceof DBSPUnaryExpression unaryRight &&
+                        unaryLeft.opcode == DBSPOpcode.WRAP_BOOL &&
+                        unaryRight.opcode == DBSPOpcode.WRAP_BOOL) {
+                    // (bool) a && (bool) b => (bool)(a && b)
+                    result = new DBSPBinaryExpression(expression.getNode(), unaryLeft.source.getType(),
+                            DBSPOpcode.AND, unaryLeft.source, unaryRight.source).wrapBoolIfNeeded();
                 }
             } else if (opcode == DBSPOpcode.OR) {
                 if (left.is(DBSPBoolLiteral.class)) {
@@ -890,6 +897,7 @@ public class Simplify extends ExpressionTranslator {
                     .appendSupplier(result::toString)
                     .newline();
         }
+        Utilities.enforce(expression.getType().sameType(result.getType()));
         super.map(expression, result);
     }
 }
