+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/visitors/outer/TemporalFilterTests.java
@@ -6,6 +6,7 @@ import org.dbsp.sqlCompiler.compiler.CompilerOptions;
 import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.sql.tools.SqlIoTest;
 import org.dbsp.sqlCompiler.compiler.visitors.inner.CanonicalForm;
+import org.dbsp.sqlCompiler.compiler.visitors.outer.temporal.BooleanExpression;
 import org.dbsp.sqlCompiler.compiler.visitors.outer.temporal.RewriteNow;
 import org.dbsp.util.Linq;
 import org.dbsp.util.Utilities;
@@ -31,7 +32,7 @@ public class TemporalFilterTests extends SqlIoTest {
                 CREATE TABLE T(a INT, b DOUBLE, c BIGINT, ts TIMESTAMP);""");
     }
 
-    List<RewriteNow.BooleanExpression> findFilters(DBSPFilterOperator operator) {
+    List<BooleanExpression> findFilters(DBSPFilterOperator operator) {
         DBSPCompiler compiler = new DBSPCompiler(new CompilerOptions());
         RewriteNow in = new RewriteNow(compiler);
         return in.findTemporalFilters(operator, operator.getClosureFunction());
